requirements.txt

google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.5
langchain-tavily>=0.1.0
langgraph>=0.0.40
pydantic>=2.0.0

tavily-python>=0.3.0
duckduckgo-search>=4.0.0


fastapi>=0.104.0
uvicorn>=0.24.0

python-dotenv>=1.0.0

streamlit>=1.28.0
